requirements1.txt
altair==4.2.0
anyio==2.2.0
appdirs==1.4.4
argcomplete==1.10.0
argon2-cffi==20.1.0
asgiref==3.3.4
async-generator==1.10
attrs==21.2.0
autopep8==1.5.7
azure-ai-formrecognizer==3.0.0
azure-ai-textanalytics==5.1.0b3
azure-cognitiveservices-vision-computervision==0.8.0
azure-cognitiveservices-vision-customvision==3.1.0
azure-common==1.1.26
azure-core==1.12.0
azure-mgmt-core==1.2.2
azure-mgmt-storage==17.0.0
azure-storage-blob==12.8.0
Babel==2.9.0
backcall==0.2.0
beautifulsoup4==4.8.0
black==21.12b0
bleach==3.3.0
blis==0.7.4
bs4==0.0.1
catalogue==2.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cryptography==3.4.7
cssselect==1.1.0
cycler==0.10.0
cymem==2.0.5
date-extractor==5.1.5
datefinder==0.7.1
dateparser==1.0.0
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.2
django==3.2
django-dotenv==1.4.2
djangorestframework==3.12.4
docx2txt==0.8
EbookLib==0.17.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl
entrypoints==0.3
et-xmlfile==1.1.0
extract-msg==0.23.1
fake-useragent==0.1.11
filelock==3.0.12
Flask==1.0.2
Flask-Cors==3.0.7
Flask-JWT-Simple==0.0.3
fuzzywuzzy==0.18.0
gitdb==4.0.9
GitPython==3.1.27
gunicorn==20.1.0
idna==2.10
image==1.5.33
IMAPClient==2.1.0
importlib-resources==5.4.0
ipykernel==5.5.0
ipython==7.22.0
ipython-genutils==0.2.0
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==3.0.2
joblib==1.0.1
JPype1==1.2.1
json-lines==0.5.0
json5==0.9.5
jsonschema==4.4.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-packaging==0.7.12
jupyter-server==1.5.1
jupyterlab==3.0.12
jupyterlab-pygments==0.1.2
jupyterlab-server==2.3.0
kiwisolver==1.3.1
lxml==4.6.2
MarkupSafe==2.0.1
matplotlib==3.4.1
mistune==0.8.4
msrest==0.6.21
murmurhash==1.0.5
mypy-extensions==0.4.3
nbclassic==0.2.6
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
nltk==3.6.2
notebook==6.3.0
numpy==1.22.3
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.1.48
openpyxl==3.0.7
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parse==1.19.0
parso==0.8.1
pathspec==0.9.0
pathy==0.4.0
PDF-Layout-Scanner==1.3.3
pdf2image==1.16.0
pdfminer.six==20201018
pdfplumber==0.5.27
phonenumbers==8.12.20
pickleshare==0.7.5
Pillow==8.2.0
pipenv==2021.5.29
platformdirs==2.4.1
pprintpp==0.4.0
preshed==3.0.5
prettytable==2.1.0
prometheus-client==0.9.0
prompt-toolkit==3.0.18
pycodestyle==2.8.0
pycparser==2.20
pycryptodome==3.10.1
pydantic==1.7.3
pyee==8.1.0
Pygments==2.8.1
PyJWT==2.0.1
pyparsing==2.4.7
PyPDF2==1.26.0
pyppeteer==0.2.5
pyquery==1.4.3
pyrsistent==0.18.0
python-dateutil==2.8.1
python-dotenv==0.19.2
python-pdfbox==0.1.8
python-pptx==0.6.18
pytz==2021.1
pywin32==300
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==22.0.3
regex==2021.4.4
requests==2.25.1
requests-html==0.10.0
requests-oauthlib==1.3.0
resume-parser==0.8
selenium==3.141.0
Send2Trash==1.5.0
Shapely==1.7.1
six==1.15.0
smart-open==3.0.0
smmap==5.0.0
sniffio==1.2.0
sortedcontainers==2.3.0
soupsieve==2.2
spacy==3.0.5
spacy-legacy==3.0.1
SpeechRecognition==3.8.1
SQLAlchemy==1.3.23
sqlparse==0.4.1
srsly==2.4.0
stemming==1.0.1
terminado==0.9.4
testpath==0.4.4
textract==1.6.3
thinc==8.0.2
tika==1.24
toml==0.10.2
tomli==1.2.3
toolz==0.11.2
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
typer==0.3.2
typing-extensions==4.0.1
tzlocal==1.5.1
Unidecode==1.2.0
urllib3==1.26.4
virtualenv==20.4.7
virtualenv-clone==0.5.4
w3lib==1.22.0
Wand==0.6.6
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
websockets==8.1
Werkzeug==0.16.0
widgetsnbextension==3.6.0
xlrd==1.2.0
XlsxWriter==1.3.7
zipp==3.7.0
